lib/oauth: add tiny library to configure different providers.
Background:
Each oauth provider has slightly different settings and qirks.
Further, at the end of the authentication, different things need to
be done in order to define the identity of the user.

Today, lib/oauth/ogoogle, lib/oauth/ogithub allow to configure
those quirks and extra steps, but the libraries need to be invoked
and used correctly by the author of the code - there are no flags/
factory method to create them.

server/astore/auth implements a couple flags to pick the tunings
correctly, and use them.

In this PR:
- add a small lib/oauth/providers to configure those providers
  via flags using a factory, like the rest of the code.

This is necessary in preparation for next commits, that will
increase the amount of configurable options per provider.

Note that it has to be a different library, as oauth cannot depend
on ogoogle and ogithub, as that would create circular dependencies.