lib/oauth: improve the interface, turn methods in free form functions.
Background:
lib/oauth originally provided only oauth as a form of authentication.

When enproxy was implemented, we added support for a "Redirector"
and "Extractor" something that would authenticate by redirecting to
a proper oauth server, and check an existing cookie rather than
perform the authentication from scratch.

The Redirector only provided a single "Authenticate" method, no
other way to eg, create a middleware or URL level authentication.
Methods like MakeAuthHandler or MakeLoginHandler remained available
exclusively for the full fledged Authenticator.

In this PR:
- add one more method in the IAuthenticator: GetCredentialsFromRequest.
  With this method added, a whole lot of methods of Authenticator
  can be turned into free functions that work with ANY IAuthenticator.
- methods as per previous point have been turned in free standing
  functions, and now work on both a full fledged Authenticator and
  Redirector.

With this PR, Redirector can almost be always used as a replacement
for Authenticator. I expect future PRs will have CLI flags and a
unified entry point over oauth to allow applications to either
use oauth, or use a redirector for login.