lib/khttp: take default port of server from environment.
In this PR:
- rather than hard code 9999 as the default port if none is specified,
  use the PORT environment variable.

This is convenient as in many cloud/virtual/containerized setups, the
port is provided via environment variables.

This reduces the distance between lib/khttp/server.go and lib/server/run.go,
which the next PRs will try to deprecate.

Tested:
added small unit test for the newly added code.