proxy/ptunnel: bugifx, reload credentials correctly at attempt time.
Background:
Authentication credentials have a time limit. Generally around 1 day.
Once they expire, they can no longer be used.

Before this PR:

if I closed my laptop today and reopened it tomorrow, an existing
tunnel or ssh session would look frozen. Not because the protocol
cannot recover, but because the tunnel/ssh is stuck using the old
credentials - even after an enkit login.

There is code in enkit to re-load the credentials, but due to the
structure of the code this was only invoked once at session establishment
time, rather than at every attempt.

For the end user, this manifested as "I closed my laptop, re-opened
the day after, sessions are dead, no matter what".

With this PR:

Correctly re-evaluate options at each re-connection attempt, including
the ones to reload the authentication cookie.

For the end user, this now works as "I closed my laptop, re-opened
the day after, run enkit login, my sessions are back in business"

Tested: this is tricky to test in an isolated test, manually verified
by forcefully casuing a disconnect (disable wifi, move to hotspot)
and checking the logs for the message indicating that the credentials
are reloaded. No reloading before PR, successful reloading after.
Also, verified that the log in a running enproxy shows the credentials
correctly.