bazel: Nuke all web frontend code, rules from orbit
This change removes all frontend code, including:

* the `//ui` dir
* associated bazel rules/macros
* loading of npm packages, frontend-related rules from WORKSPACE

The catalyst for this is INFRA-4192 - use of these rules uses a
`managed_directory` feature of WORKSPACE that is removed in bazel 6.0
and is thus blocking bazel updates. Fixing the UI code is:

* probably time-consuming, to understand the nuances of JS/TS
  compilation, by looking at https://docs.aspect.build/guides/rules_js_migration/
* of little value, since none of this code is actually used in anything

The only dependency that actually exists is on some font and CSS assets;
these are inlined into the repo directly, with a note on how to update
them if needed.

This also has ancillary benefits:

* reduced build times, since node modules don't need to be fetched
* no build-out-of-sync-type errors that necessitated a `bazel clean
  --expunge && rm -rf ui/node_modules to resolve

Tested: no

* TODO: should test that asset replacement doesn't cause issues

Jira: INFRA-4192