bes_publisher: Publish more messages
This change adds more messages to the list published by `bes_publisher`,
along with a test that ensures the attributes are aggregated correctly.

This change also adds a generic map library, with convenience functions
for maps so that these simple functions no longer need to be duplicated
in the codebase.

Tested: unit tests