bazel/utils/messaging.bzl: fix bug in fileowners.
Turns out `fileowner` as written does not work, bazel complains
that the package variable is used before assignment.

This PR fixes it by simply renaming the variable so not to conflict
with the macro name.

This probably slipped through as the use of the macro was
dropped, and a package macro was added.