bazel/utils: add a cc_dynamic_wrapper rule for LD_LIBRARY_PATH.
Problem:
When using cc_foreign_rules and dynamic binaries, the generated
binary is often not runnable as is: the linker cannot find the library
at run time due to things like rpath, bazel sandboxing, or the
generated directory names.

So far we've worked around this on a need by need basis, by hard
coding paths in shell or python scripts to export the correct
LD_LIBRARY_PATH.

This PR:
defines a cc_dynamic_wrapper rule that generates a shell wrapper
for dynamic binaries that sets the LD_LIBRARY_PATH correctly
before running the binary.

This rule also embeds some convenience in updating the runfiles
definition with all the specified dependencies.

Read the .bzl file for documentation and details on how to use
the rule itself.