third_party: Add Python proto rules
This change adds `py_proto_library` rules for the Bazel BES proto, so
that code in internal can depend on it.

Since internal only wants to use the top-level BES proto, only that rule
needs to have public visibility; the dependencies can stick with
restricted visibility.

Tested: Code in internal that requires the proto builds with this change
when `--override_repository` is set appropriately