bazel: Update to 6.0
This change updates enkit to bazel 6.0.0.

An incompatible flag has to be added in order to allow for this - this
flag can be removed once dependencies are updated (likely involving
migrating current protobuf rules to rules_proto).

`remote_run_test` behavior appears to change, needing to copy libgtest
in addition to the other files, which seems harmless so the goldens are
updated accordingly.

`astore` is changed to ensure the build does not include cgo, as
otherwise its usage in internal causes a segfault during a cgo call. It
is not clear why this becomes an issue with the bazel 6.0 update.

Tested:
* TODO: Test to make sure this doesn't break internal in an obvious way

Jira: INFRA-4384