auth server: add support for propagating groups. (#799)
* lib/oauth: support multiple verifiers.

Background:
At the end of the oauth process, once a token has been obtained,
a verifier is invoked. This verifier checks the claims of the token,
and creates an identity for the user.

Today, a single verifier can be specified. Generally, the verifier
makes a call to the provider API to fetch eg, the email or information
of the user, or uses a key/certificate to extract and verify it from
the token.

This PR is meant to introduce no functional changes, just juggles
the code around. Specifically:
- there is now a "Verifier" interface, that defines 1) the scopes
  the oauth token needs for the verifier to work 2) provides a
  function to actually verify the token.
- moves the code around in the ogoogle and ogithub libraries to
  implement the verifier interface. Note that the code is exactly
  the same, todo by todo and comment by comment,  just moved in
  the interface.
- changes factory.go to use multiple verifiers.

Also, adds some defense in depth so if verifiers succeed but return
no valid credentials, the authentication fails.

Tested:
Server started with dev credentials after the change successfully
authenticates.

* lib/oauth: add support for propagating groups as part of identity.

* lib/oauth: retrieve group membership and store it in google provider.

* astore/server/auth: propagate groups as certificate principals.

* lib/oauth: add tiny library to configure different providers.

Background:
Each oauth provider has slightly different settings and qirks.
Further, at the end of the authentication, different things need to
be done in order to define the identity of the user.

Today, lib/oauth/ogoogle, lib/oauth/ogithub allow to configure
those quirks and extra steps, but the libraries need to be invoked
and used correctly by the author of the code - there are no flags/
factory method to create them.

server/astore/auth implements a couple flags to pick the tunings
correctly, and use them.

In this PR:
- add a small lib/oauth/providers to configure those providers
  via flags using a factory, like the rest of the code.

This is necessary in preparation for next commits, that will
increase the amount of configurable options per provider.

Note that it has to be a different library, as oauth cannot depend
on ogoogle and ogithub, as that would create circular dependencies.

* astore/server: migrate it to use the new lib/oauth/providers library.

This allows to deprecate the code in credentials.go, and some other
minor simplifications in the server.

* lib/oauth: make ogoogle provider configurable from CLI.

* lib/oauth/providers: add group renamer and keeper.

In this PR:
- added a verifier to prune and rename groups returned by
  a provider. This is helpful when a) not all groups a user is member
  of should be stored in its authentication cookie and b) when the name
  of those groups needs to be mangled.

* lib/oauth: allow for optional verifiers, mark groups as optional by default.

Background:
We know the cloudidentity API is fragile. If an error happens with that API
today, the user is not allowed to log in.

In this PR:
- create a Verifier that allows to mark any other verifier to be optional.
- allow the groups API to be marked optional. If optional, errors are ignored.
- mark the groups API optional by default.

* lib/logger: add variable with default Go logger.

This just saves some typing, can find this same pattern in many
places around the code.

Also, improved a comment describing the function.

* lib/oauth (and astore): create and propagate a logger.

In this PR:
- create and propagate a logger object to the oauth library,
  so events can be logged.
- introduce some initial logging in the Verify API (more to come).

* lib/oauth: run go fmt after all changes.

* lib/oauth: make Warnf an Errorf if a verifier returns an error.

* lib/oauth/ogoogle: use an "enum" rather than two booleans.

... as recommended by Scott in code review.

* lib/oauth/ogoogle: add unit test covering groups and optional.

This is by no means an end to end test, but covers at least
the functionality of the optional marker and of the groups
fetcher.