astore_download: Use prebuilt astore client
This change switches to using a pre-built `astore` client for
`astore_download` and `astore_upload` rules, since building the client
is not supported in certain environments currently.

The astore client was built manually and uploaded to astore at
`bazel/workspace_deps/astore`, before being publicly exposed so that it
could be fetched via HTTP (fetching via astore would cause a bit of a
chicken-and-egg scenario).

Tested:

Jira: INFRA-4783