enproxy: Define config changes for conditionally adding headers
This change defines the new types and additions to the enproxy
config struct that enable mapping headers to request based on the user's
groups.

The proposal enables a config spelled like:

```
- from:
    host: 'monitoring.corp.enfabrica.net'
  to: http://127.0.0.1:3000/
  transform:
    xwebauth: "set"
    mapRequestHeadersByGroup:
    - header: X-Webauth-Role
      groupMapping:
      - group: foo
        value: Admin
      - group: bar
        value: Editor
      - group: ""
        value: Viewer
```

To translate to: "Add `X-Webauth-Role: Admin` to users in the `foo`
group; add `X-Webauth-Role: Editor` to users in the `bar` group; add
`X-Webauth-Role: Viewer` to everyone else".

Tested: N/A

Jira: INFRA-4919