enproxy: Implement mapRequestHeadersByGroup
This change implements the "map request headers by group" functionality,
which allows header values to be assigned to a given header
conditionally based on group matching.

In order to simplify the logic a bit, a helper library is created to
convert slices into maps, reducing the need for nested `for` loops
within the feature implementation itself.

Tested: not yet

Jira: INFRA-4919