bazel/linux: various feature additions in preparation for next PR (#878)
* bazel/linux: remove nv_.* related rules, that are unused and obsolete.

* bazel/linux: add support for initialization steps in vm setup.

Background:
Until now, a VM could only be configured to run a pre-defined set
of commands, but there was no distinction between commands necessary
to setup the VM, and commands necessary to eg, run a test.

This PR introduces a way to provide test initializations.

* bazel/linux: expand bundle capabilities to support list of binaries.

This is mostly a maintenance/cleanup commit. A few (related) changes here:

- All macros/rules that need to compute commands from a bundle are
  now using a single, common, macro returning a struct.
- This one macro does not rely on the verbose parameter anymore.
  Instead, it adds source information whenever it has not already
  been provided. This requires a new parameter in the provider,
  origin, which is added.
- The bundle definition is changed to support a list of commands/binaries,
  rather than just one.

This is in preparation of next PRs.

* bazel/linux: expand bundles to support recursive expansion.

Before this commit:
A bundle could only reference a binary in _bin, a set of commands
in _cmd, and a sinlge set of args in _args.

After this commit:
- _bin, _cmds, _args are unmodified for backward compatibility.
- a new attribute is introduced: setup, init, run, cleanup, check.
  This new attribute can take a **list** of things to run.
  The "thing to run" can be a binary, or **another bundle**.

This means that now we can create bundles of bundles, and it should
all work as expected.

Additionally:
- fixed cfg = "" for targets running in the VM! They were compiled
  for the exec platform, rather than the target platform.

* bazel/linux: improve qemu_test target to propagate init correctly.

This commit changes the qemu_test target so that setup steps
are propagated in the bundle correctly, and marked as init steps.