go: Update ecosystem
This change updates the golang ecosystem, including:

* rules_go to latest release
* gazelle to latest release
* Go SDK to latest release (1.20.3)

This was hopefully going to fix an issue with gazelle and strict
dependencies in internal, but doesn't - however, it is still less risky
to be on the latest versions if there are no negative effects.

One of the unit tests testing byte slices is flaky for unknown reasons
(test can't decide if it wants strings in the golden cases or not) so
the comparer func is modified to normalize both the "got" and "want"
slices to sidestep this issue.

Tested:
* `bazel test //...` passes
* `bazel test //infra/cj/... --override_repository=enkit=<path to
enkit>` works in internal