bazel/linux: fix breakages introduced by PR#878, variuos improvements (#896)
* bazel/linux: propagate arguments correctly.

Background:
Some tests and qemu bundles define a binary to run, and arguments to use for that binary.

My previous PR broke the propagation of arguments when a binary was listed explicitly in
the bundle. This commit fixes it.

* bazel/linux: use a better approach to handle recursion, fix bugs.

Background:
Some targets accept vm_bundles as input. vm_bundles define sets of
commands to run. Sets of commands can be represented in two ways:
- as a list of string, the literal commands to run + runfiles.
- as a RuntimeInfo provider, maintaining information about binaries,
  targets, runfiles, ...

Some parts of the testing and building framework required processing
bundles recursively. The current code did so by flattening all
RuntimeInfo into a list of commands + runfiles, and propagating it
throughout.

This makes it impossible to 1) distinguish debug commands (eg to
output the target name) from the actual command run, 2) distinguish
which commands belong to one test or another, or which commands belong
to one target or another - which ended up being a big problem.

In this commit:
- use the other approach described above: propagate RuntimeInfo
  instead of lists of commands. So the information is not lost.
- expand RuntimeInfo at the end of the processing, not as part
  of the expansion steps.

Tested:
All unit tests in the internal/ repository that started failing
with my previous PR are now succeeding.