bazel: Make all patches repo-absolute
Currently, patch labels are specified starting with `//` (omitting
`@enkit`) which makes them "repo-relative": when another repo calls a
macro that has a repository rule that references a patch with this
format, it looks for the patch within itself instead of enkit.

To avoid having to duplicate these patches in other repos, this change
makes the patch paths absolute, so other repos will look for them
within enkit itself.

Tested:
* Built internal targets via `--override_repository`; confirmed this
  fixes build in internal