bb_reporter: Add no-op server
This change adds an implementation of a `CompletedActionLogger` server
that basically drops all messages, while:
* logging a single message to stdout every once in a while, so we can
  see some actual test data
* collecting metrics, so we can get an idea of the scale of the data

Eventually this service should upload messages to some sort of database,
but that will come in a future PR. This PR is sufficient to confirm that
the implementation of the skeleton is sufficient to interoperate with
Buildbarn.

Tested: Server builds and starts.

Jira: INFRA-5725