bb_reporter: Parse to BigQuery struct
This change parses the `CompletedAction` message into a struct which
will represent a BigQuery table entry. The proto message is not used
directly in case it has too much nesting to be a useable BigQuery
schema.

Pushing to BigQuery is still not implemented, but the parsed struct is
printed periodically on stdout instead of the received message.

This change also modifies the batch size parameter, boosting that and
the batch timeout params to flags in the process.

Tested: Deployed; successfully showing entries and no log
errors/warnings

Jira: INFRA-5726