bb_reporter: Write results to BigQuery (#905)
This change implements BigQuery writing for each batch of `ActionRecord`
instances.

Each struct is taught to implement `bigquery.ValueSaver`, which is an
interface whereby each is converted to a row. This is a bit extra for a
few reasons:
* the bigquery library can infer a schema from a struct instance (though
  it is not clear how the field names are inferred)
* there is already a conversion from proto message to struct, and now a
  second conversion from struct to map - it's possible to do this in one
  conversion. This double-conversion approach is chosen because it makes
  the metrics/logging more explicit (rather than needing to parse the
  Bigquery client library errors).

The BigQuery insertion code implements an interface that looks like
internal's `memo.Inserter`, in case the latter is ever backported to
this repo. A second variation is introduced (`BatchInserter`) with
similar semantics, but for a list of values instead.

Tested: builds

Jira: INFRA-5725