ssh agent: add timeout, flags, improve error handling. (#893)
* lib/kauth: solve enkit login bug.

Background:
I had some reports of folks being unable to load certificates
with `enkit login`, but never pinpointed it. I attributed it to
older enkit version, or some environment issues, as `enkit machine upgrade`
or `enkit machine setup` would normally solve it.

Turns out that a `defer agent.Close()` is actually responsible
for killing the ssh agent as soon as it's started, when started
from enkit login. This works just fine if folks have `enkit agent run`
in their .bashrc, which is the recommended setting at enfabrica.

* lib/kflags: add support for array of strings passed as arguments.

* lib/kflags: run go fmt in the directory.

... and no other changes in this commit.

* proxy/ptunnel/commands/agent: drop deprecated string, noop.

Just removes a global variable that's no longer used anywhere.
(Code was modified to use printf with no template instead)

* lib/kcerts: improve error messages and comments.

This change is a noop: just improves the error messages returned with
some use of multierror and fmt.Errorf() + a few comments added, and
minor stylistic changes.

* lib/kcerts: go fmt

This cleans up random style violations in the directory, in preparation
for next commit.

* lib/kcerts: add configurable timeout in handling the ssh agent.

In this commit:
- use a timeout when talking with the ssh agent.
- make the timeout configurable, plumb flags propagation throughout.

* lib/kcerts: separate object state from serialized state.

Background:
SSHAgent objects have state that has to be serialized to disk and
read back. Additionally, SSHAgent objects have attributes that are
necessary for the object to work (for example: a Close() callback).

In this commit:
- separate state from attributes.
- add a unit test to ensure everything is still working as expected.

(this is mostly in preparation of next commit)

* dependencies: update testify library to the latest version.

... it contains a lot more helper function, including helpers for
error handlers.

* lib/kcerts: add more ssh agent flags, plumb flags throughout, add tests.

This commit closes the series:
- uses / plumbs the new flags throughout, following the same conventions
  used in the rest of the repository.
- improves in the error management and reporting as well.
- adds a unit test verifying that the new flags are effective, and that
  the added logic to verify agent functionality and timeouts is working
  as expected.

* lib/kcerts: use composition instead of embedding for SSHAgentState.

As per comment in https://github.com/enfabrica/enkit/pull/893#discussion_r1178380969,
to make code more readable.