bb_shim: Extend to take a map of bytestream hosts
The buildbuddy shim rewrites bytestream URLs in query strings to trick
BuildBuddy into fetching content from Buildbarn. Prior to this change, a
shim can only rewrite any host it sees to a single bytestream host.

With this change, the shim will take a map of rewrites to perform, and
rewrite the bytestream host depending on what the original bytestream
hsot was. This will allow us to instruct a single BuildBuddy instance to
redirect to multiple Buildbarn instances, as long as the hosts in the
original bytestream requests are distinct in some way. (currently, they
are distinct by either port number if a tunnel is used, or URL if no
tunnel is used).

This code can be improved in many ways, but this change just extends
functionality and modifies tests to test the new functionality.

Tested: unit tests only