bazel/dependencies/abseil: patch abseil to support multiple flag libraries.
In the latest abseil release there is no sane way to use multiple flag
libraries: there is no easy way to tell abseil to 1) ignore the flags
it does not know about and 2) LEAVE THEM in the returned argv 3) strip
the flags it does parse and know about. 2 out of 3 are possible using
the internal API, 3 is not.

This patch adds the ability to ignore flags abseil does not know about
AND leave them in the returned argv, which enables us to use multiple
flag libraries for code that is not based on abseil.

abseil in head provides a different function to achieve a similar
result, but it's very clumsy to use as it CHANGES the order of argv
arguments, which affects other flag parsing libraries.

Tested:
See corresponding PR in internal using and testing the feature.