bestie: Fix error loop; add fetch metrics
This change adds a metric that records HTTP success/failure + status
code for each bytestream fetch.

While looking at this metric, it became clear that all fetches are
returning an HTTP 200 even when the item is not found, and the stream is
closing early. This might hypothetically cause any wrapper on the
`io.Reader` to error. In the case of zip files, we append this error to
a list in a hot loop and retry the next file, which could cause an OoM
if this codepath always returns a non-EOF error. To guard against this,
the code exits the loop on any zip iteration failure.

Tested: Built, deployed, waiting for failures.

Jira: INFRA-6331