vscode: Set up Go language server
The Go language server / bazel integration finally looks stable enough
to use - this change proposes that we make the setup canonical, and adds
some supporting scripts to make this possible.

This PR adds:
* the recommended gopackagesdriver.sh script from the editor setup
  guide: https://github.com/bazelbuild/rules_go/wiki/Editor-setup#2-launcher-script
* the recommended editor config from the editor setup guide:
  https://github.com/bazelbuild/rules_go/wiki/Editor-setup#2-launcher-script

Since editor-specific config files are gitignore'd, this change removes
the gitignore for vscode settings. The rationale here is that some of
these settings are truly project-specific (`gopls` LSP configuration
for bazel actually is different from normal Go projects, and is
different per bazel workspace) and thus belong here.

Note: `gopls` completion is broken unless VSCode is run in the container
context, due to path issues. This means either exec'ing vscode from the
container, or attaching to the container from a host-launched VSCode
instance.