gazelle: Fix pruning of necessary dependencies (#932)
There are currently some repositories fetched in the gazelle-generated
list from go.mod that are needed, but don't actually have imports in the
Go source. This leads to an issue where Go tooling is constantly trying
to prune them from go.mod, and so Gazelle is trying to prune them from
go_repositories.

Up until now, we have all been carefully stepping around this issue.
This change fixes the issue by adding a dummy package with a few imports
that ensures those packages are kept around.