bazel/utils: fix exec_test to use the target platform.
Tests in bazel are normally built for the target platform, not
for the exec (or host) platform. The target platform is what needs to be
tested (although there can be multiple target platforms).

Further, cfg = "exec" applies to all direct and indirect
dependencies of a rule, as it changes what the target is.

Basically, any exec_test causes the entire dependency graph
of the test to be rebuilt for a different platform.