googletest: Update to 1.13.0
This change updates googletest to 1.13.0, to pick up fixes to the bazel
rule/macro glue within that make it more compatible with newer versions
of bazel.

Updating googletest causes tests in this repo to fail, unless flags are
passed to set the C++ standard higher by default. We'd prefer to match
the flags passed in internal (`-std=c++20`), but those don't seem to
work in this context, so we pass the next most approximate flag
(`-std=c++2a`).

Tested: `bazel test //...` in this repo