gee setup: add "enable_bcompare" (#945)
gee setup: add "enable_bcompare"

Add `gee setup enable_bcompare` to configure gee to use the commercial
BeyondCompare tool to resolve merge conflicts.  The goal here is to make
it easy for Enfabricators to evaluate the use of BeyondCompare to see if
buying a license is worthwhile.

Tested: Manually tested in my hw-dev container.