rules_foreign_cc: Update to latest version
This change updates `rules_foreign_cc` to the latest released version.

The new version takes a function that we were using private, so this
change adds a patch to re-export it. `meson.bzl` is updated where some
of the function signatures/contracts changed, and meson is updated to
the latest release as well.

This change will need to be coupled with #<internal PR> when updating
enkit in internal, as there is a change to how make variables are
written/substituted during expansion that breaks existing builds
otherwise.

Tested: See #<internal PR> in internal

Jira: INFRA-6722