astore_package: Add attributes to allow for patching (#949)
Currently, we can apply patches on top of `http_archive` tars, but not
`astore_package` tars. This change adds `patches` and `patch_args`
attributes, and calls the same underlying `patch` util function that
implements this for `http_archive`.

In theory, we can simply apply patches before uploading tars to astore.
However, this inserts manual steps in the mirroring process, adding
friction and increasing the probability of error. With this change, we
can mirror code to astore as-is, and patch after-the-fact, which is at
least enforced by automation.