enproxy: Don't include port number in hostname matching
It would seem that if clients include the port number in requests to
enproxy, the port number is including in the host matching logic, which
will result in 404s. This is inconvenient, and will result in duplicate
mappings/more complex wildcard matching to ensure all the possible
spellings of a particular hostname are matched correctly.

This change proposes that the port number be stripped from requests
before they are matched.

---

This PR is currently just a unit test modification only, not to be
merged, to illustrate and verify that this is indeed an issue.

Jira: INFRA-6906