`enkit outputs mount`: Repurpose for local bb_clientd
Prior to this change, `enkit outputs mount` attempted to start and
manage a long-running bb_clientd process that talks via tunnels to
buildbarn and buildbuddy. With this change, `enkit outputs mount` no
longer manages `bb_clientd`; a long-running `bb_clientd` instance must
be pre-configured on the host or in the development environment.

This allows for the deletion of much of the flaky code, which is a net
improvement. Some changes are made to how the linking works:
* `unmount` now deletes the hardlink forest for a particular invocation
  ID, which will free disk resources.
* Because multiple users could be using a particular invocation ID,
  hardlink forests are created on a per-user per-invocation ID basis
* `shutdown` deletes all hardlink forests for that particular user

The code is still haphazardly organized and can be rewritten and
reorganized into packages to reduce incidental code complexity, but this
can be done in a future PR.

Tested: manually on dev machine (TODO: insert steps)