bazel/linux: Explicitly list all kernel tree outputs
Listing only directories as outputs of the kernel tree rule
causes the "dependency checking of directories is unsound"
warning and can fail under some scenarios.  One example is when
the tree is used as an input of a genrule.

See https://github.com/bazelbuild/bazel/issues/1415#issuecomment-482490890
as well as https://bazel.build/reference/be/general#filegroup

We can explicitly specify all the files in the kernel tree because
it's a repository rule.