rules_foreign_cc: Update to latest master commit
rules_foreign_cc was recently updated to the latest release, though this
release is now a year old and doesn't contain the newest `meson` rule,
which we can use to drop our current implementation. This PR updates
`rules_foreign_cc` to the latest master commit.

Tested: `@qemu-uds` targets still build in internal with this change
(and tweaks to make `meson()` rules compatible with rules_foreign_cc)