rules_python: Update to 0.24.0; add patch to handle circular deps
This change:

* updates `rules_python` to the latest released version (which is just
  generally a good thing to do, as it reduces tech debt)
* adds a patch that allows for specific packages to be patched to avoid
  circular dependencies

The patch comes from
https://github.com/bazelbuild/rules_python/pull/1166#issuecomment-1641878169
and is necessary as airflow has made multiple of their packages
dependent on each other on purpose for reasons that are unlikely to be
walked back on. This is apparently valid in pip-land, but of course not
valid in bazel. The patch is a nice medium-term workaround while the
community figures out exactly how this should be handled.

Tested:
* `bazel test //...` in this repo (needed to regenerate
  `requirements.in`)
* affected-targets detection in internal to verify that this does solve
  the apache-airflow circular deps issue causing the tool to break

Not tested:
* rules_python update in internal

Jira: INFRA-7317